finished factories,seeders

+++ database/migrations/2023_09_19_056000_create_products_table.php
@@ -1,39 +1,53 @@
 <?php
 
+use App\Models\Brand;
+use App\Models\Category;
+use App\Models\Producer;
+use App\Models\Product;
 use Illuminate\Database\{Migrations\Migration, Schema\Blueprint};
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
+
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
+
             $table->id();
             $table->timestamps();
-            $table->string('name')->unique();
-            $table->string('image')->nullable();
-            $table->string('bread_crumbs')->nullable();
-            $table->string('characteristics')->nullable();
-            $table->text('describe')->nullable();
-            $table->text('options')->nullable();
-            $table->integer('rating')->nullable();
-            $table->unsignedInteger('estimates')->nullable();
-            $table->unsignedInteger('count');
-            $table->unsignedFloat('price');
-            $table->unsignedFloat('old_price')->nullable();
-            $table->foreignId('category_id')->constrained()->cascadeOnDelete()->cascadeOnUpdate();
-            $table->foreignId('producer_id')->constrained()->cascadeOnDelete()->cascadeOnUpdate();
+            $table->string('slug');
+            $table->string('title')->unique();
+            $table->string('thumbnail')->nullable();
+            $table->foreignIdFor(Brand::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
+            $table->foreignIdFor(Producer::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
+            $table->unsignedInteger('price')->default(0);
+            $table->unsignedInteger('old_price')->nullable();
+
+
+            // $table->string('bread_crumbs')->nullable();
+            // $table->text('characteristics')->nullable();
+            // $table->text('describe')->nullable();
+            // $table->text('options')->nullable();
+            // $table->unsignedFloat('rating')->nullable();
+            // $table->unsignedInteger('count_estimates');
+            // $table->unsignedInteger('estimate')->nullable();
+        });
+
+        Schema::create('category_product', function (Blueprint $table) {
+            $table->id();
+
+            $table->foreignIdFor(Category::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
+            $table->foreignIdFor(Product::class)->constrained()->cascadeOnDelete()->cascadeOnUpdate();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
+
     public function down(): void
     {
-        Schema::dropIfExists('products');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('category_product');
+            Schema::dropIfExists('products');
+        }
     }
 };
